+++ src/write_buffer.rs
@@ -30,6 +30,7 @@
 
 use crate::client::Client;
 use crate::constants;
+use crate::db_type::DB_TYPE_CURSOR;
 use crate::db_type::DbType;
 
 pub struct WriteBuffer {
@@ -417,6 +418,23 @@ where
             <T>::to_buf(value, buf, db_type, write_length);
         } else {
             <T>::default().to_buf_null(buf);
+        }
+    }
+}
+
+impl ToBuf for &'static DbType {
+    fn to_buf(
+        &self,
+        buf: &mut WriteBuffer,
+        _db_type: &'static DbType,
+        _write_length: bool,
+    ) {
+        match *self {
+            &DB_TYPE_CURSOR => {
+                buf.write_u8(1); // length of integer
+                buf.write_u8(0); // value of integer (cursor number)
+            }
+            _ => buf.write_u8(0),
         }
     }
 }